+++ include/entidy/View.h
@@ -11,7 +11,7 @@ namespace entidy
 {
 using namespace std;
 
-using Entity = size_t;
+using Entity = uint32_t;
 
 class IndexerImpl;
